refactor(rust): extract query + aggregate into db/ (slice 5/9)
Move find-style and aggregation query logic out of the lib.rs god-module into dedicated submodules:

- db/query.rs: execute_mql_query_impl, count_documents_impl, explain_mql_query_impl
- db/aggregate.rs: execute_aggregate_impl, explain_aggregate_query_impl

The thin #[tauri::command] wrappers stay in lib.rs and call the impls via crate-root re-exports (pub use db::query::{...}, db::aggregate::{...}), so command registration and the test suite are unchanged. Pure move; no behavior change. cargo test: 60 passed.
