+++ src-tauri/src/lib.rs
@@ -19,7 +19,9 @@ mod vault;
 mod db;
 mod state;
 mod window;
+pub use db::aggregate::{execute_aggregate_impl, explain_aggregate_query_impl};
 pub use db::gridfs::{download_gridfs_file_impl, list_gridfs_files_impl, GridFsFileInfo};
+pub use db::query::{count_documents_impl, execute_mql_query_impl, explain_mql_query_impl};
 pub use db::schema::{analyze_schema_impl, infer_schema, FieldStat, SchemaReport, TypeCount};
 pub use db::version::get_mongodb_version_impl;
 pub use state::{AppState, LockExt};
@@ -688,162 +690,6 @@ pub async fn list_indexes_impl(
     Ok(indexes)
 }
 
-pub async fn execute_mql_query_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-    sort: &str,
-    projection: &str,
-    limit: i64,
-    skip: i64,
-) -> Result<Vec<String>, String> {
-    // Validate projection JSON up front (applies on the real path; mock ignores fields).
-    let projection_doc: Option<mongodb::bson::Document> =
-        if projection.trim().is_empty() || projection.trim() == "{}" {
-            None
-        } else {
-            let val: serde_json::Value = serde_json::from_str(projection)
-                .map_err(|e| format!("Invalid MQL projection JSON: {}", e))?;
-            Some(
-                mongodb::bson::to_document(&val)
-                    .map_err(|e| format!("BSON conversion error: {}", e))?,
-            )
-        };
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return mock_db::execute_mock_query(database, collection, filter, sort, limit, skip);
-    }
-
-    let filter_val: serde_json::Value = if filter.trim().is_empty() {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(filter).map_err(|e| format!("Invalid MQL filter JSON: {}", e))?
-    };
-
-    let sort_val: serde_json::Value = if sort.trim().is_empty() {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(sort).map_err(|e| format!("Invalid MQL sort JSON: {}", e))?
-    };
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let db = client.database(database);
-    let coll = db.collection::<mongodb::bson::Document>(collection);
-
-    // Convert filter serde_json::Value to BSON Document
-    let filter_doc = mongodb::bson::to_document(&filter_val)
-        .map_err(|e| format!("BSON conversion error: {}", e))?;
-
-    // Convert sort serde_json::Value to BSON Document
-    let sort_doc: Option<mongodb::bson::Document> =
-        if sort_val.is_object() && !sort_val.as_object().unwrap().is_empty() {
-            let doc = mongodb::bson::to_document(&sort_val)
-                .map_err(|e| format!("BSON conversion error: {}", e))?;
-            Some(doc)
-        } else {
-            None
-        };
-
-    let mut find_builder = coll.find(filter_doc);
-    if let Some(sort) = sort_doc {
-        find_builder = find_builder.sort(sort);
-    }
-    if let Some(projection) = projection_doc {
-        find_builder = find_builder.projection(projection);
-    }
-    if limit > 0 {
-        find_builder = find_builder.limit(limit);
-    }
-    if skip > 0 {
-        find_builder = find_builder.skip(skip as u64);
-    }
-
-    let mut cursor = find_builder
-        .await
-        .map_err(|e| format!("Query failed: {}", e))?;
-
-    let mut results = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = cursor.next().await {
-        let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-        let json_val: serde_json::Value =
-            serde_json::to_value(&doc).map_err(|e| format!("BSON to JSON error: {}", e))?;
-        results.push(serde_json::to_string(&json_val).unwrap());
-    }
-
-    Ok(results)
-}
-
-pub async fn count_documents_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-) -> Result<u64, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return mock_db::count_mock_documents(database, collection, filter);
-    }
-
-    let trimmed = filter.trim();
-    let is_empty_filter = trimmed.is_empty() || trimmed == "{}";
-
-    let filter_val: serde_json::Value = if is_empty_filter {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(filter).map_err(|e| format!("Invalid MQL filter JSON: {}", e))?
-    };
-    let filter_doc = mongodb::bson::to_document(&filter_val)
-        .map_err(|e| format!("BSON conversion error: {}", e))?;
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-
-    // Empty filter: use the fast metadata estimate instead of a full scan.
-    let count = if is_empty_filter {
-        coll.estimated_document_count()
-            .await
-            .map_err(|e| format!("Failed to estimate document count: {}", e))?
-    } else {
-        coll.count_documents(filter_doc)
-            .await
-            .map_err(|e| format!("Count failed: {}", e))?
-    };
-    Ok(count)
-}
-
 fn csv_escape(raw: &str) -> String {
     if raw.contains(',') || raw.contains('"') || raw.contains('\n') || raw.contains('\r') {
         format!("\"{}\"", raw.replace('"', "\"\""))
@@ -1205,192 +1051,6 @@ pub async fn start_collection_export_impl(
     });
 
     Ok(task)
-}
-
-pub async fn execute_aggregate_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    pipeline: &str,
-) -> Result<Vec<String>, String> {
-    // Parse and validate the pipeline (a JSON array of stage documents) up front so a
-    // malformed pipeline fails clearly regardless of the connection type.
-    let pipeline_val: serde_json::Value = if pipeline.trim().is_empty() {
-        serde_json::Value::Array(Vec::new())
-    } else {
-        serde_json::from_str(pipeline)
-            .map_err(|e| format!("Invalid aggregation pipeline JSON: {}", e))?
-    };
-    let stages_val = pipeline_val
-        .as_array()
-        .ok_or_else(|| "Aggregation pipeline must be a JSON array of stages".to_string())?;
-    let mut stages: Vec<mongodb::bson::Document> = Vec::with_capacity(stages_val.len());
-    for stage in stages_val {
-        let doc = mongodb::bson::to_document(stage)
-            .map_err(|e| format!("Invalid aggregation stage: {}", e))?;
-        stages.push(doc);
-    }
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Err("Aggregation pipelines are not supported on mock connections".to_string());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-
-    let mut cursor = coll
-        .aggregate(stages)
-        .await
-        .map_err(|e| format!("Aggregation failed: {}", e))?;
-
-    let mut results = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = cursor.next().await {
-        let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-        let json_val: serde_json::Value =
-            serde_json::to_value(&doc).map_err(|e| format!("BSON to JSON error: {}", e))?;
-        results.push(serde_json::to_string(&json_val).unwrap());
-    }
-
-    Ok(results)
-}
-
-/// Explain an entire aggregation pipeline (M1), not just its `$match` stage.
-/// Mirrors `execute_aggregate_impl`'s validation and is real-connection-only.
-pub async fn explain_aggregate_query_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    pipeline: &str,
-) -> Result<String, String> {
-    // Parse and validate the pipeline (a JSON array of stage documents) up front.
-    let pipeline_val: serde_json::Value = if pipeline.trim().is_empty() {
-        serde_json::Value::Array(Vec::new())
-    } else {
-        serde_json::from_str(pipeline)
-            .map_err(|e| format!("Invalid aggregation pipeline JSON: {}", e))?
-    };
-    let stages_val = pipeline_val
-        .as_array()
-        .ok_or_else(|| "Aggregation pipeline must be a JSON array of stages".to_string())?;
-    let mut stages: Vec<mongodb::bson::Document> = Vec::with_capacity(stages_val.len());
-    for stage in stages_val {
-        let doc = mongodb::bson::to_document(stage)
-            .map_err(|e| format!("Invalid aggregation stage: {}", e))?;
-        stages.push(doc);
-    }
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Err("Aggregation explain is not supported on mock connections".to_string());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let db = client.database(database);
-    let command = mongodb::bson::doc! {
-        "explain": {
-            "aggregate": collection,
-            "pipeline": stages,
-            "cursor": {}
-        },
-        "verbosity": "executionStats"
-    };
-
-    let explain_result = db
-        .run_command(command)
-        .await
-        .map_err(|e| format!("Explain failed: {}", e))?;
-
-    let json_val: serde_json::Value =
-        serde_json::to_value(&explain_result).map_err(|e| format!("BSON to JSON error: {}", e))?;
-
-    Ok(serde_json::to_string_pretty(&json_val).unwrap())
-}
-
-pub async fn explain_mql_query_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-) -> Result<String, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Ok(mock_db::get_mock_explain(database, collection, filter));
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let db = client.database(database);
-    let filter_val: serde_json::Value = if filter.trim().is_empty() {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(filter).map_err(|e| format!("Invalid MQL filter JSON: {}", e))?
-    };
-
-    let filter_doc = mongodb::bson::to_document(&filter_val)
-        .map_err(|e| format!("BSON conversion error: {}", e))?;
-
-    let command = mongodb::bson::doc! {
-        "explain": {
-            "find": collection,
-            "filter": filter_doc
-        },
-        "verbosity": "executionStats"
-    };
-
-    let explain_result = db
-        .run_command(command)
-        .await
-        .map_err(|e| format!("Explain failed: {}", e))?;
-
-    let json_val: serde_json::Value =
-        serde_json::to_value(&explain_result).map_err(|e| format!("BSON to JSON error: {}", e))?;
-
-    Ok(serde_json::to_string_pretty(&json_val).unwrap())
 }
 
 pub async fn create_index_impl(
